+++ src/irvine/oeis/a001/A001226.java
@@ -15,6 +15,6 @@ public class A001226 extends Sequence1 {
   @Override
   public Z next() {
     mN += 2;
-    return Z.ONE.shiftLeft((int) Euler.phiAsLong(mN)).subtract(1).divide(mN);
+    return Z.ONE.shiftLeft(Euler.phiAsLong(mN)).subtract(1).divide(mN);
   }
 }
