+++ src/irvine/oeis/a006/A006023.java
@@ -49,7 +49,7 @@ public class A006023 extends Sequence0 {
           prod = RING.multiply(prod, RING.pow(RING.oneMinusXToTheN(k), v[k]));
         }
       }
-      final Q coeff = new Q(Z.ONE.shiftLeft((int) getPowersOfTwo(v)), SymmetricGroup.per(v));
+      final Q coeff = new Q(Z.ONE.shiftLeft(getPowersOfTwo(v)), SymmetricGroup.per(v));
       prod = RING.multiply(prod, coeff);
       sum = RING.add(sum, prod);
     }
