+++ src/irvine/oeis/a016/A016047.java
@@ -17,6 +17,6 @@ public class A016047 implements Sequence {
   @Override
   public Z next() {
     mP = mPrime.nextPrime(mP);
-    return Jaguar.factor(Z.ONE.shiftLeft((int) mP).subtract(1)).leastPrimeFactor();
+    return Jaguar.factor(Z.ONE.shiftLeft(mP).subtract(1)).leastPrimeFactor();
   }
 }
